wip: add nessus e2e test
